♻️ Refactor package to split `core` and `spec`
We've decided to keep `WandbFile` in `core` and move both `WandbArtifactStore` and `WandbFileSystem` to `spec`, but defining a base class named `WandbBaseFileSystem` with the shared stuff
